Signature verification on JavaScript side finally working
It might be necessary to adjust how the req.body is read. On user.js it has been necessary to make body a string and then parse to JSON. The body was a buffer of bytes. This happened because the android app is sending data with content type "application/octet-stream" instead of "application/json".